+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 * Update API to 5.11.0, tools to 5.5.1 and other packages
 * Create custom styles for keyboard focus of aster slices
 * Split d3 into submodules
-* Remove interactivity-utils
 * Fix context menu for center label, outer lines and legends
+* Remove interactivity-utils
 
 ## 1.7.1
 * Fix scaling when outerline is enabled